Addressbook Splitted in Extra Module
