--- translations/de.json
@@ -1,4 +1,0 @@
-{
-  "title": "Eingehender Anruf",
-  "noCall": "Keine kürzlichen Anrufe."
-}
