--- translations/en.json
@@ -1,4 +1,0 @@
-{
-  "title": "Incoming Call",
-  "noCall": "No recent calls."
-}
